+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# [2.3.0]
+- Fix bug zoom out.
+- Adds `animateZoom` method in `Camera`.
+- Experimental State manager. Example [here](https://github.com/RafaelBarbosatec/bonfire/issues/218#issuecomment-1058121200)
+
 # [2.2.5]
 - Fix crash mentioned in th issue [#255](https://github.com/RafaelBarbosatec/bonfire/issues/225)
 - Fix crash mentioned in th issue [#227](https://github.com/RafaelBarbosatec/bonfire/issues/227)